refactor common tests
