+++ skore/tests/unit/sklearn/comparison/cross_validation/test_report.py
@@ -1,17 +1,7 @@
-import re
-from io import BytesIO
-
-import joblib
 import pytest
-from sklearn.datasets import make_classification, make_regression
+from sklearn.datasets import make_regression
 from sklearn.dummy import DummyClassifier, DummyRegressor
 from skore import ComparisonReport, CrossValidationReport
-
-
-@pytest.fixture
-def classification_data():
-    X, y = make_classification(class_sep=0.1, random_state=42)
-    return X, y
 
 
 @pytest.fixture
@@ -21,18 +11,6 @@ def cv_report_classification(classification_data):
         DummyClassifier(strategy="uniform", random_state=0), X, y
     )
     return cv_report
-
-
-@pytest.fixture
-def cv_reports(classification_data):
-    X, y = classification_data
-    cv_report_1 = CrossValidationReport(
-        DummyClassifier(strategy="uniform", random_state=0), X, y
-    )
-    cv_report_2 = CrossValidationReport(
-        DummyClassifier(strategy="uniform", random_state=1), X, y
-    )
-    return cv_report_1, cv_report_2
 
 
 @pytest.fixture
@@ -79,79 +57,19 @@ def test_non_string_report_names(cv_reports):
         ComparisonReport({0: cv_report_1, "1": cv_report_2})
 
 
-@pytest.fixture
-def report(cv_reports):
-    cv_report_1, cv_report_2 = cv_reports
-    return ComparisonReport([cv_report_1, cv_report_2])
-
-
-def test_help(capsys, report):
-    """Check the help menu works."""
-    report.help()
-
-    captured = capsys.readouterr()
-    assert "Tools to compare estimators" in captured.out
-
-    # Check that we have a line with accuracy and the arrow associated with it
-    assert re.search(
-        r"\.accuracy\([^)]*\).*\(↗︎\).*-.*accuracy", captured.out, re.MULTILINE
-    )
-
-
-def test_repr(report):
-    """Check the `__repr__` works."""
-
-    assert "ComparisonReport" in repr(report)
-
-
-def test_metrics_repr(report):
-    """Check the repr method of `report.metrics`."""
-    repr_str = repr(report.metrics)
-    assert "skore.ComparisonReport.metrics" in repr_str
-    assert "help()" in repr_str
-
-
-def test_pickle(tmp_path, report):
-    """Check that we can pickle a comparison report."""
-    with BytesIO() as stream:
-        joblib.dump(report, stream)
-        joblib.load(stream)
-
-
-def test_cross_validation_report_cleaned_up(report):
-    """
-    When a CrossValidationReport is passed to a ComparisonReport, and computations are
-    done on the ComparisonReport, the CrossValidationReport should remain pickle-able.
-
-    Non-regression test for bug found in:
-    https://github.com/probabl-ai/skore/pull/1512
-    """
-    report.metrics.report_metrics()
-
-    with BytesIO() as stream:
-        joblib.dump(report.reports_[0], stream)
-
-
-def test_metrics_help(capsys, report):
-    """Check that the help method writes to the console."""
-    report.metrics.help()
-    captured = capsys.readouterr()
-    assert "Available metrics methods" in captured.out
-
-
 @pytest.mark.parametrize("data_source", ["train", "test", "X_y"])
-def test_get_predictions(report, classification_data, data_source):
+def test_get_predictions(report_cv_reports, classification_data, data_source):
     if data_source == "X_y":
         X, _ = classification_data
-        predictions = report.get_predictions(
+        predictions = report_cv_reports.get_predictions(
             X=X, data_source=data_source, response_method="predict"
         )
 
     else:
-        predictions = report.get_predictions(
+        predictions = report_cv_reports.get_predictions(
             data_source=data_source, response_method="predict"
         )
 
-    assert len(predictions) == len(report.reports_)
-    for i, cv_report in enumerate(report.reports_):
+    assert len(predictions) == len(report_cv_reports.reports_)
+    for i, cv_report in enumerate(report_cv_reports.reports_):
         assert len(predictions[i]) == cv_report._cv_splitter.n_splits
